Add storage verb
